adding merged useNFT hook
