+++ src/hooks/useNFT.ts
@@ -1,14 +1,15 @@
 import { useContext } from 'react';
-import useSWR from 'swr';
 
 import { NFTFetchContext } from '../context/NFTFetchContext';
-import { OpenseaNFTDataType } from '../fetcher/AuctionInfoTypes';
-import { transformOpenseaResponse } from '../fetcher/OpenseaUtils';
+import { NFTDataType } from '../fetcher/AuctionInfoTypes';
+import { useOpenseaNFT } from './useOpenseaNFT';
+import { ZORA_MEDIA_CONTRACT_BY_NETWORK } from '../constants/addresses';
+import { useZNFT } from './useZNFT';
 
 export type useNFTType = {
   currencyLoaded: boolean;
   error?: string;
-  data?: OpenseaNFTDataType;
+  data?: NFTDataType;
 };
 
 type OptionsType = {
@@ -20,7 +21,7 @@ type OptionsType = {
 /**
  * Fetches on-chain NFT data and pricing for the given zNFT id
  *
- * @param contractAddress address of the contract
+ * @param contractAddress address of the contract, if null and tokenID is passed in, a ZNFT is assumed
  * @param tokenId id of NFT to fetch blockchain information for
  * @param options SWR flags and an option to load currency info
  * @returns useNFTType hook results include loading, error, and chainNFT data.
@@ -31,44 +32,27 @@ export function useNFT(
   options: OptionsType = {}
 ): useNFTType {
   const fetcher = useContext(NFTFetchContext);
-  const { loadCurrencyInfo = false, refreshInterval, initialData } = options || {};
 
-  const nftData = useSWR(
-    contractAddress && tokenId ? ['loadGenericNFT', contractAddress, tokenId] : null,
-    (_, contractAddress, tokenId) => fetcher.loadNFTDataUntransformed(contractAddress, tokenId),
-    { dedupingInterval: 0 }
-  );
-  const auctionData = useSWR(
-    contractAddress && tokenId ? ['loadAuctionForNFT', contractAddress, tokenId] : null,
-    (_, contractAddress, tokenId) => fetcher.loadAuctionInfo(contractAddress, tokenId)
-  );
-
-  const nftResponseData = nftData.data as any;
-  const currencyData = useSWR(
-    nftResponseData && loadCurrencyInfo
-      ? ['loadCurrencies', auctionData.data?.auctionCurrency]
-      : null,
-    (_, ...currencies) => fetcher.loadCurrencies(currencies),
-    {
-      refreshInterval,
-      dedupingInterval: 0,
-    }
-  );
-
-  let data: OpenseaNFTDataType | undefined = undefined;
-  if (nftData.data !== undefined) {
-    data = transformOpenseaResponse(
-      nftData.data,
-      auctionData.data,
-      currencyData.data,
-    );
-  } else {
-    data = initialData;
+  if (!contractAddress) {
+    contractAddress = ZORA_MEDIA_CONTRACT_BY_NETWORK[fetcher.networkId];
   }
 
+  const isZoraContractAddress =
+    contractAddress === ZORA_MEDIA_CONTRACT_BY_NETWORK[fetcher.networkId];
+
+  const nonZoraNFT = useOpenseaNFT(
+    !isZoraContractAddress ? contractAddress : undefined,
+    !isZoraContractAddress ? tokenId : undefined,
+    options
+  );
+
+  const zoraNFT = useZNFT(isZoraContractAddress ? tokenId : undefined, options);
+
+  let data = nonZoraNFT || zoraNFT;
+
   return {
-    currencyLoaded: !!currencyData.data,
-    error: nftData.error,
-    data,
+    currencyLoaded: !!data.currencyLoaded,
+    error: data.error,
+    data: data.data,
   };
 }
